# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Game Review API
  version: 1.0.0
  description: CRUD API for games and reviews.
servers:
  - url: http://localhost:8080
paths:
  /games:
    get:
      summary: List games
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
    post:
      summary: Create a game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /games/{id}:
    get:
      summary: Get a game
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Not found
    put:
      summary: Update a game
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Not found
    delete:
      summary: Delete a game
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
  /reviews:
    get:
      summary: List reviews
      parameters:
        - in: query
          name: gameId
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
    post:
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid rating
        '404':
          description: Game not found
  /reviews/{id}:
    get:
      summary: Get a review
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Not found
    put:
      summary: Update a review
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid rating
        '404':
          description: Not found
    delete:
      summary: Delete a review
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
components:
  parameters:
    GameId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
    ReviewId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        genre:
          type: string
        releaseYear:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - genre
        - releaseYear
        - createdAt
        - updatedAt
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        gameId:
          type: integer
          format: int64
        reviewer:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - gameId
        - reviewer
        - rating
        - comment
        - createdAt
        - updatedAt
    CreateGameRequest:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        releaseYear:
          type: integer
      required:
        - title
        - genre
        - releaseYear
    UpdateGameRequest:
      type: object
      properties:
        title:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        releaseYear:
          type: integer
          nullable: true
    CreateReviewRequest:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
        reviewer:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
      required:
        - gameId
        - reviewer
        - rating
        - comment
    UpdateReviewRequest:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
          nullable: true
        reviewer:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
        comment:
          type: string
          nullable: true